Refactor portfolio into shared modules for maintainability
- Extract shared Framer Motion variants into src/lib/animations.js
- Centralize all site content and data into src/data/siteData.js and src/data/navItems.js
- Create reusable SectionWrapper, SectionHeader, and BackToTopButton components
- Simplify App.jsx and NavBar.jsx by removing duplicated markup and using shared modules
- Fix pre-existing ESLint false positive for motion import in varsIgnorePattern
