+++ src/App.jsx
@@ -7,44 +7,17 @@ import LanguageSkills from "./components/LanguageSkills";
 import GitHubStats from "./components/GitHubStats";
 import LeetCodeStats from "./components/LeetCodeStats";
 import Contact from "./components/Contact";
-import { useState, useEffect, useMemo } from "react";
-import { motion } from "framer-motion";
+import SectionWrapper from "./components/SectionWrapper";
+import BackToTopButton from "./components/BackToTopButton";
+import { useMemo } from "react";
 
 const VALID_PASSWORD = "password123";
 
 function App() {
-  const [showBackToTop, setShowBackToTop] = useState(false);
-
   const isAuthenticated = useMemo(() => {
     const params = new URLSearchParams(window.location.search);
     return params.get("password") === VALID_PASSWORD;
   }, []);
-
-  useEffect(() => {
-    const handleScroll = () => {
-      // Show button when scrolled down more than 300px
-      const shouldShow = window.scrollY > 300;
-      setShowBackToTop(shouldShow);
-    };
-
-    // Initial check
-    handleScroll();
-
-    // Add scroll listener
-    window.addEventListener("scroll", handleScroll, { passive: true });
-
-    return () => {
-      window.removeEventListener("scroll", handleScroll);
-    };
-  }, []);
-
-  const scrollToTop = () => {
-    // Find the home section and scroll to it
-    const homeSection = document.getElementById("home");
-    if (homeSection) {
-      homeSection.scrollIntoView({ behavior: "smooth" });
-    }
-  };
 
   const handleNavClick = (e) => {
     e.preventDefault();
@@ -55,118 +28,36 @@ function App() {
     }
   };
 
-  const fadeInUp = {
-    hidden: { opacity: 0, y: 60 },
-    visible: {
-      opacity: 1,
-      y: 0,
-      transition: {
-        duration: 0.6,
-        ease: "easeOut",
-      },
-    },
-  };
-
   return (
     <div className="relative">
       <NavBar onNavClick={handleNavClick} isAuthenticated={isAuthenticated} />
       <main className="pt-16">
-        <motion.div
-          id="home"
-          className="min-h-screen"
-          initial="hidden"
-          whileInView="visible"
-          viewport={{ once: true, margin: "-100px" }}
-          variants={fadeInUp}
-        >
+        <SectionWrapper id="home">
           <Landing />
-        </motion.div>
-        <motion.div
-          id="about"
-          className="min-h-screen"
-          initial="hidden"
-          whileInView="visible"
-          viewport={{ once: true, margin: "-100px" }}
-          variants={fadeInUp}
-        >
+        </SectionWrapper>
+        <SectionWrapper id="about">
           <AboutMe />
-        </motion.div>
+        </SectionWrapper>
         {isAuthenticated && (
-          <motion.div
-            id="experience"
-            className="min-h-screen"
-            initial="hidden"
-            whileInView="visible"
-            viewport={{ once: true, margin: "-100px" }}
-            variants={fadeInUp}
-          >
+          <SectionWrapper id="experience">
             <Experience />
-          </motion.div>
+          </SectionWrapper>
         )}
-        <motion.div
-          id="skills"
-          className="min-h-screen"
-          initial="hidden"
-          whileInView="visible"
-          viewport={{ once: true, margin: "-100px" }}
-          variants={fadeInUp}
-        >
+        <SectionWrapper id="skills">
           <LanguageSkills />
-        </motion.div>
-        <motion.div
-          id="github"
-          className="min-h-screen"
-          initial="hidden"
-          whileInView="visible"
-          viewport={{ once: true, margin: "-100px" }}
-          variants={fadeInUp}
-        >
+        </SectionWrapper>
+        <SectionWrapper id="github">
           <GitHubStats />
-        </motion.div>
-        <motion.div
-          id="leetcode"
-          className="min-h-screen"
-          initial="hidden"
-          whileInView="visible"
-          viewport={{ once: true, margin: "-100px" }}
-          variants={fadeInUp}
-        >
+        </SectionWrapper>
+        <SectionWrapper id="leetcode">
           <LeetCodeStats />
-        </motion.div>
-        <motion.div
-          id="contact"
-          className="min-h-screen"
-          initial="hidden"
-          whileInView="visible"
-          viewport={{ once: true, margin: "-100px" }}
-          variants={fadeInUp}
-        >
+        </SectionWrapper>
+        <SectionWrapper id="contact">
           <Contact />
-        </motion.div>
+        </SectionWrapper>
       </main>
 
-      <button
-        onClick={scrollToTop}
-        className={`fixed bottom-8 right-8 z-[100] transition-all duration-300 ${
-          showBackToTop ? "opacity-100" : "opacity-0 pointer-events-none"
-        } w-[50px] h-[50px] rounded-full bg-base-300 border-2 border-primary/20 font-semibold flex items-center justify-center shadow-lg cursor-pointer overflow-hidden hover:w-[140px] hover:rounded-[50px] group`}
-        aria-label="Back to top"
-      >
-        <div className="relative w-full h-full flex items-center justify-center">
-          <svg
-            className="w-3 transition-transform duration-300 group-hover:translate-y-[-200%]"
-            viewBox="0 0 384 512"
-          >
-            <path
-              d="M214.6 41.4c-12.5-12.5-32.8-12.5-45.3 0l-160 160c-12.5 12.5-12.5 32.8 0 45.3s32.8 12.5 45.3 0L160 141.2V448c0 17.7 14.3 32 32 32s32-14.3 32-32V141.2L329.4 246.6c12.5 12.5 32.8 12.5 45.3 0s12.5-32.8 0-45.3l-160-160z"
-              className="fill-base-content"
-            />
-          </svg>
-          <span className="absolute text-base-content text-[0px] opacity-0 transition-all duration-300 group-hover:text-[13px] group-hover:opacity-100">
-            Back to Top
-          </span>
-        </div>
-      </button>
+      <BackToTopButton />
     </div>
   );
 }
