+++ src/components/LeetCodeStats.jsx
@@ -1,37 +1,15 @@
 import React from "react";
 import { motion } from "framer-motion";
+import { itemVariants } from "../lib/animations";
+import { leetcodeUsername } from "../data/siteData";
+import SectionHeader from "./SectionHeader";
 
 const LeetCodeStats = () => {
-  const itemVariants = {
-    hidden: { opacity: 0, y: 20 },
-    visible: {
-      opacity: 1,
-      y: 0,
-      transition: {
-        duration: 0.6,
-        ease: "easeOut",
-      },
-    },
-  };
-
-  const username = "Nikky0510";
-
   return (
     <section className="min-h-screen py-16 bg-base-300 flex items-center">
       <div className="container mx-auto px-4">
         <div className="max-w-4xl mx-auto">
-          <motion.div
-            className="text-center mb-12"
-            initial={{ opacity: 0, y: 20 }}
-            whileInView={{ opacity: 1, y: 0 }}
-            viewport={{ once: true, margin: "-100px" }}
-            transition={{ duration: 0.6, ease: "easeOut" }}
-          >
-            <h1 className="text-5xl font-bold mb-4 bg-gradient-to-r from-primary to-secondary bg-clip-text text-transparent">
-              Problem Solving
-            </h1>
-            <div className="w-24 h-1 bg-gradient-to-r from-primary to-secondary mx-auto rounded-full"></div>
-          </motion.div>
+          <SectionHeader title="Problem Solving" />
 
           <motion.div
             initial="hidden"
@@ -45,7 +23,7 @@ const LeetCodeStats = () => {
                 <h2 className="card-title text-2xl mb-4 text-primary text-center mx-auto">LeetCode Stats</h2>
                 <div className="rounded-lg overflow-hidden">
                   <img
-                    src={`https://leetcard.jacoblin.cool/${username}?theme=dark&font=Baloo%202&ext=contest`}
+                    src={`https://leetcard.jacoblin.cool/${leetcodeUsername}?theme=dark&font=Baloo%202&ext=contest`}
                     alt="LeetCode Stats"
                     className="w-full"
                   />
