+++ src/data/navItems.js
@@ -1,0 +1,9 @@
+export const navItems = [
+  { href: "#home", label: "Home" },
+  { href: "#about", label: "About" },
+  { href: "#experience", label: "Experience", authRequired: true },
+  { href: "#skills", label: "Skills" },
+  { href: "#github", label: "GitHub" },
+  { href: "#leetcode", label: "LeetCode" },
+  { href: "#contact", label: "Contact" },
+];
