+++ src/data/siteData.js
@@ -1,0 +1,95 @@
+export const languages = [
+  {
+    name: "Vue",
+    logo: "https://cdn.jsdelivr.net/gh/devicons/devicon/icons/vuejs/vuejs-original.svg",
+  },
+  {
+    name: "React",
+    logo: "https://cdn.jsdelivr.net/gh/devicons/devicon/icons/react/react-original.svg",
+  },
+  {
+    name: "HTML",
+    logo: "https://cdn.jsdelivr.net/gh/devicons/devicon/icons/html5/html5-original.svg",
+  },
+  {
+    name: "JavaScript",
+    logo: "https://cdn.jsdelivr.net/gh/devicons/devicon/icons/javascript/javascript-original.svg",
+  },
+  {
+    name: "Spring Boot",
+    logo: "https://cdn.jsdelivr.net/gh/devicons/devicon/icons/spring/spring-original.svg",
+  },
+  {
+    name: "FastAPI",
+    logo: "https://cdn.jsdelivr.net/gh/devicons/devicon/icons/fastapi/fastapi-original.svg",
+  },
+  {
+    name: "Java",
+    logo: "https://cdn.jsdelivr.net/gh/devicons/devicon/icons/java/java-original.svg",
+  },
+  {
+    name: "Python",
+    logo: "https://cdn.jsdelivr.net/gh/devicons/devicon/icons/python/python-original.svg",
+  },
+  {
+    name: "Grafana",
+    logo: "https://cdn.jsdelivr.net/gh/devicons/devicon/icons/grafana/grafana-original.svg",
+  },
+  {
+    name: "InfluxDB",
+    logo: "https://cdn.jsdelivr.net/gh/devicons/devicon/icons/influxdb/influxdb-original.svg",
+  },
+];
+
+export const tools = [
+  {
+    name: "Git",
+    logo: "https://cdn.jsdelivr.net/gh/devicons/devicon/icons/git/git-original.svg",
+  },
+  {
+    name: "Docker",
+    logo: "https://cdn.jsdelivr.net/gh/devicons/devicon/icons/docker/docker-original.svg",
+  },
+  {
+    name: "Kubernetes",
+    logo: "https://cdn.jsdelivr.net/gh/devicons/devicon/icons/kubernetes/kubernetes-plain.svg",
+  },
+  {
+    name: "Terraform",
+    logo: "https://cdn.jsdelivr.net/gh/devicons/devicon/icons/terraform/terraform-original.svg",
+  },
+];
+
+export const experiences = [
+  {
+    title: "Job Title",
+    company: "Company Name",
+    period: "Start Date – End Date",
+    description:
+      "Placeholder description of role and responsibilities. Replace with actual experience details.",
+  },
+  {
+    title: "Job Title",
+    company: "Company Name",
+    period: "Start Date – End Date",
+    description:
+      "Placeholder description of role and responsibilities. Replace with actual experience details.",
+  },
+  {
+    title: "Job Title",
+    company: "Company Name",
+    period: "Start Date – End Date",
+    description:
+      "Placeholder description of role and responsibilities. Replace with actual experience details.",
+  },
+];
+
+export const contactInfo = {
+  phone: "+1 (551) 358-3376",
+  email: "nikkyjsoriano@gmail.com",
+  location: "East Brunswick, NJ",
+};
+
+export const githubUsername = "nikkyjsoriano";
+
+export const leetcodeUsername = "Nikky0510";
